fixed factor product implementation

+++ src/in/ac/iitb/aml/model/CliquePotential.java
@@ -12,23 +12,23 @@ import java.util.Map;
  */
 public class CliquePotential {
 	public CliquePotential(BitSet cliqueNodes,
-			Map<BitSet, Float> potential) {
+			Map<BitSet, Double> potential) {
 		super();
 		this.cliqueNodes = cliqueNodes;
 		this.potential = potential;
 	}
 	private BitSet cliqueNodes;
-	private Map<BitSet,Float> potential;
+	private Map<BitSet,Double> potential;
 	public BitSet getCliqueNodes() {
 		return cliqueNodes;
 	}
 	public void setCliqueNodes(BitSet cliqueNodes) {
 		this.cliqueNodes = cliqueNodes;
 	}
-	public Map<BitSet, Float> getPotential() {
+	public Map<BitSet, Double> getPotential() {
 		return potential;
 	}
-	public void setPotential(Map<BitSet, Float> potential) {
+	public void setPotential(Map<BitSet, Double> potential) {
 		this.potential = potential;
 	}
